# apps/api/docs/swagger.yaml
definitions:
  github_com_projectx_api_internal_platform_httpx.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
info:
  contact:
    name: Project-X support
  description: |-
    REST API for Project-X — the AI-run marketing + studio operations platform.
    Auth is a JWT stored in an HTTP-only cookie (set by POST /api/v1/auth/login).
    When calling protected endpoints from a tool like Swagger UI or curl, log in first and reuse the returned cookie.
  license:
    name: Proprietary
  title: Project-X API
  version: "1.0"
paths:
...
  /api/v1/studios/{studioId}/messaging/channels/telegram-web/backfill:
    get:
      description: Returns the current backfill status (e.g. running, done, failed,
        none) and imported message count for a studio's Telegram Web session, read
        from our database.
      parameters:
      - description: Studio ID
        in: path
        name: studioId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get Telegram Web chat history backfill status
      tags:
      - Messaging - Telegram Web
    post:
      description: 'Kicks off a one-time chat-history import for a QR-linked Telegram
        Web session: marks the backfill status "running" in our database, then proxies
        the trigger request to the tg-web Node service. Fire-and-forget from the admin''s
        perspective — progress is polled via tgWebBackfillStatus.'
      parameters:
      - description: Studio ID
        in: path
        name: studioId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Trigger Telegram Web chat history backfill
      tags:
      - Messaging - Telegram Web
  /api/v1/studios/{studioId}/messaging/channels/telegram-web/disconnect:
    post:
      description: Marks the studio's Telegram Web channel as disconnected in our
        database, then proxies the disconnect request to the tg-web Node service to
        tear down the underlying MTProto session.
      parameters:
      - description: Studio ID
        in: path
        name: studioId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Disconnect a studio's Telegram Web session
      tags:
      - Messaging - Telegram Web
  /api/v1/studios/{studioId}/messaging/channels/telegram-web/password:
    post:
      consumes:
      - application/json
      description: Forwards the studio's Telegram 2FA (two-step verification) password,
        submitted during QR login, to the tg-web Node service to complete the MTProto
        sign-in.
      parameters:
      - description: Studio ID
        in: path
        name: studioId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: could not read request body
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "502":
          description: tg-web service unavailable
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Submit 2FA password for a Telegram Web login
      tags:
      - Messaging - Telegram Web
  /api/v1/studios/{studioId}/messaging/channels/telegram-web/qr:
    get:
      description: Proxies to the tg-web Node service to fetch the current QR code
        (or pairing status) for a studio's Telegram personal-account (MTProto) session.
      parameters:
      - description: Studio ID
        in: path
        name: studioId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get Telegram Web QR login code
      tags:
      - Messaging - Telegram Web
  /api/v1/studios/{studioId}/messaging/channels/telegram-web/status:
    get:
      description: Proxies to the tg-web Node service to fetch the current connection
        status (e.g. connected, disconnected, awaiting scan/password) of a studio's
        Telegram Web session.
      parameters:
      - description: Studio ID
        in: path
        name: studioId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get a studio's Telegram Web session status
      tags:
      - Messaging - Telegram Web
...
  /internal/tg-web/backfill:
    post:
      consumes:
      - application/json
      description: Receives a batch of historical messages for one chat, imported
        by the tg-web Node service after a QR-linked session connects. Called repeatedly,
        once per chat (or in chat-sized pages), not all-at-once; empty messages are
        skipped. Internal-only (Docker-network only, not exposed via nginx) — no session/API-key
        auth, authenticated via the shared `x-internal-key` header instead.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body or studioId
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      summary: Import a batch of Telegram Web backfill messages
      tags:
      - Messaging - Internal
  /internal/tg-web/backfill-done:
    post:
      consumes:
      - application/json
      description: Lets the tg-web Node service report that it has finished walking
        every chat, so the admin UI can stop showing "running". On genuine success
        (not failed), publishes a shared "QR-linked backfill finished" event to kick
        off post-backfill AI summarization of imported conversations. Internal-only
        (Docker-network only, not exposed via nginx) — no session/API-key auth, authenticated
        via the shared `x-internal-key` header instead.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body or studioId
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      summary: Mark Telegram Web backfill as finished
      tags:
      - Messaging - Internal
  /internal/tg-web/backfill-running:
    post:
      consumes:
      - application/json
      description: Lets the tg-web Node service mark history import as started the
        moment a session connects, so the admin UI reflects an import already underway
        rather than sitting on "none". Internal-only (Docker-network only, not exposed
        via nginx) — no session/API-key auth, authenticated via the shared `x-internal-key`
        header instead.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body or studioId
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      summary: Mark Telegram Web backfill as running
      tags:
      - Messaging - Internal
  /internal/tg-web/connected:
    post:
      consumes:
      - application/json
      description: Called by the tg-web Node service when a studio's MTProto session
        successfully signs in. Upserts the Telegram Web channel with the paired phone/username
        and encrypted session string. Internal-only (Docker-network only, not exposed
        via nginx) — no session/API-key auth, authenticated via the shared `x-internal-key`
        header instead.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body or studioId
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projectx_api_internal_platform_httpx.ErrorResponse'
      summary: Report a Telegram Web session connected
      tags:
      - Messaging - Internal
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    description: JWT session cookie set by POST /api/v1/auth/login (cookie name configurable
      via COOKIE_NAME).
    in: cookie
    name: px_session
    type: apiKey
swagger: "2.0"